# internal/contrato/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "notarum",
    "version": "1.0.0",
    "description": "API abierta de sólo lectura del Boletín Oficial de la República Argentina. Entrega el calendario de ediciones, cada edición por sección y fecha con sus avisos, cada aviso con su texto, los anexos en PDF y el catálogo de rubros. No opina, no clasifica, no puntúa: entrega el dato crudo y bien tipado.\n\nSin clave para leer. Límite de pedidos por IP. Los datos son del Boletín Oficial; notarum es sólo una caché legible.",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "contact": {
      "name": "notarum",
      "url": "https://github.com/diegoparras/notarum"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "esta instancia"
    }
  ],
...
  "paths": {
...
    "/v1/nacional": {
      "get": {
        "tags": [
          "nacional"
        ],
        "summary": "Buscar normativa nacional",
        "description": "Busca en las 428 mil normas de InfoLEG por título y materias. A diferencia de /v1/buscar, que recorre las ediciones del Boletín día por día, acá no hace falta saber la fecha. Buscar el tipo y el número —\"ley 24240\"— trae esa norma primero. Necesita NOTARUM_BUSCADOR_INFOLEG=1 en la instancia.",
        "operationId": "buscarNacional",
        "parameters": [
          {
            "name": "texto",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Palabras a buscar. No hace falta poner los acentos ni los puntos de los números."
          },
          {
            "name": "tipo",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Ley, Decreto, Resolución… La lista está en /v1/nacional/tipos."
          },
          {
            "name": "desde",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Año de sanción más viejo."
          },
          {
            "name": "hasta",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Año de sanción más nuevo."
          },
          {
            "name": "con_texto",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Con 1 deja sólo las que tienen el texto publicado."
          },
          {
            "name": "pagina",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limite",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 30
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Las normas que cumplen.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultadoNacional"
                }
              }
            }
          },
          "400": {
            "description": "Un año mal escrito o un rango al revés.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "El buscador está apagado en esta instancia.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "El catálogo de InfoLEG todavía no se sincronizó.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/nacional/{id}": {
      "get": {
        "tags": [
          "nacional"
        ],
        "summary": "Una norma nacional",
        "description": "Trae una norma entera por su identificador de InfoLEG, con sus fechas, su organismo y cuántas normas la modificaron.",
        "operationId": "verNormaNacional",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "example": 638
          }
        ],
        "responses": {
          "200": {
            "description": "La norma.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NormaNacional"
                }
              }
            }
          },
          "400": {
            "description": "El identificador no es un número.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No está en el catálogo guardado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
...
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error",
          "origen"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "detalle": {
            "type": "string"
          },
          "origen": {
            "type": "string",
            "enum": [
              "sitio",
              "notarum",
              "pedido"
            ],
            "description": "De quién es la culpa: del Boletín Oficial, de notarum, o de cómo se armó el pedido."
          },
          "sin_edicion": {
            "type": "boolean",
            "description": "Ese día no hubo edición."
          }
        }
      },
...
      "NormaNacional": {
        "type": "object",
        "description": "Una norma nacional de InfoLEG.",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Identificador de InfoLEG.",
            "example": 638
          },
          "tipo": {
            "type": "string",
            "example": "Ley"
          },
          "numero": {
            "type": "string",
            "example": "24240"
          },
          "titulo": {
            "type": "string"
          },
          "fecha_sancion": {
            "type": "string",
            "format": "date"
          },
          "tiene_texto": {
            "type": "boolean",
            "description": "Si InfoLEG publicó el texto. Lo tienen 192.707 de las 428.380."
          },
          "ficha": {
            "type": "string",
            "description": "La página de la norma en InfoLEG."
          },
          "texto": {
            "type": "string",
            "description": "El texto original, cuando está publicado."
          },
          "en_notarum": {
            "type": "string",
            "description": "La ruta de esta API que trae la norma entera."
          }
        }
      },
      "ResultadoNacional": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer"
          },
          "pagina": {
            "type": "integer"
          },
          "normas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NormaNacional"
            }
          },
          "hay_mas": {
            "type": "boolean"
          }
        }
      }
    }
  }
}